+++ Vault-Coder/Changes.md
@@ -5,6 +5,7 @@
 - `"description_link"` in the model structure document is now considered to refer to a class property and not a column
   - This may cause different results in cases where custom table names were used alongside the `"described"` property
 ### New Features
+- Added **ManyXAccessLike** -trait writing in cases where a class contains combinations and when specifically requested
 - You can now write property length limit rules, which will be collected to a separate document (see README for details)
 - You can now customize column, table, property, etc. naming rules in the model declaration document 
   (see README for more details)